workaround for checking CVC. Details below.
Stripe test card that will fail with a CVC error: 4000000000000127

Problem 1: Stripe will let us create a payment token using the CVC but will not validate the CVC. To see this, look for the "Demo" section on this page: https://stripe.com/docs/custom-form
Use the test card above, and you will notice that a token is returned with no error.

Problem 2: The PledgeFormStep2 calls the newPledge action using the token, apparently assuming that the card has been validated. The newPledge action is a crazy rabbit hole, and it doesn't correctly return the error message to the reducer. So the reducer doesn't get error details, and it previously returned an empty string instead of the actual message. As far as the form was concerned that meant there was an error- so it didn't close the form, but it had no error text to display.

For now, I'm just assuming that the error adding a pledge was because of a CC or CVC or EXP typo, and telling the user that.